Created sortandsearch.py
My assignment was to create a menu-driven console app in Python that creates a list of random integers, then uses this list with the following options: Interpolation search, Bogo Sort (That was my assignment) and Heap Sort. 
On each sort you have to enter a step variable as input, that will show the sorted list after the number of steps you gave.